requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
streamlit==1.45.0
requests==2.32.3
python-dotenv==1.1.0
sentence-transformers==4.1.0
chromadb==0.6.3
litellm==1.68.1
langchain==0.3.25
langchain-community==0.3.24
langchain-huggingface==0.2.0
langchain-chroma==0.2.3
pymupdf==1.25.5
python-docx==1.1.2
markdown2==2.5.3
Pillow==11.2.1
httpx==0.28.1
numpy==2.2.5
typing-extensions==4.13.2
PyYAML==6.0.2
python-multipart==0.0.20
boto3>=1.28.57
